+++ tests/test_bake_project.py
@@ -113,105 +113,26 @@ def test_bake_with_defaults(cookies):
         assert "python_boilerplate" in found_toplevel_files
         assert "tox.ini" in found_toplevel_files
         assert "tests" in found_toplevel_files
-
-        mkdocs_yml = os.path.join(result._project_dir, "mkdocs.yml")
-        with open(mkdocs_yml, "r") as f:
-            lines = f.readlines()
-            assert "  - authors: authors.md\n" in lines
-
-
-def test_bake_without_author_file(cookies):
-    with bake_in_temp_dir(cookies, extra_context={"create_author_file": "n"}) as result:
-        found_toplevel_files = [f.basename for f in result.project.listdir()]
-        assert "AUTHORS.md" not in found_toplevel_files
-        doc_files = [f.basename for f in result.project.join("docs").listdir()]
-        assert "authors.md" not in doc_files
-
-        # make sure '-authors: authors.md' not appeared in mkdocs.yml
-        mkdocs_yml = os.path.join(result._project_dir, "mkdocs.yml")
-        with open(mkdocs_yml, "r") as f:
-            lines = f.readlines()
-            assert "  - authors: authors.md\n" not in lines
-
-
-@pytest.mark.parametrize(
-    "license_info",
-    [
-        ("MIT", "MIT "),
-        (
-            "BSD-3-Clause",
-            "Redistributions of source code must retain the "
-            + "above copyright notice, this",
-        ),
-        ("ISC", "ISC License"),
-        ("Apache-2.0", "Licensed under the Apache License, Version 2.0"),
-        ("GPL-3.0-only", "GNU GENERAL PUBLIC LICENSE"),
-    ],
-)
-def test_bake_selecting_license(cookies, license_info):
-    license, target_string = license_info
-    with bake_in_temp_dir(
-        cookies, extra_context={"open_source_license": license}
-    ) as result:
-        assert target_string in result.project.join("LICENSE").read()
-        assert license in result.project.join(_DEPENDENCY_FILE).read()
-
-
-def test_bake_not_open_source(cookies):
-    with bake_in_temp_dir(
-        cookies, extra_context={"open_source_license": "Not open source"}
-    ) as result:
-        found_toplevel_files = [f.basename for f in result.project.listdir()]
         assert _DEPENDENCY_FILE in found_toplevel_files
-        assert "LICENSE" not in found_toplevel_files
-        assert "License" not in result.project.join("README.md").read()
-        assert "license" not in result.project.join(_DEPENDENCY_FILE).read()
+        assert "LICENSE" in found_toplevel_files
+        assert "License" in result.project.join("README.md").read()
+        assert "license" in result.project.join(_DEPENDENCY_FILE).read()
 
-
-def test_not_using_pytest(cookies):
-    with bake_in_temp_dir(cookies, extra_context={"use_pytest": "n"}) as result:
-        assert result.project.isdir()
-        # Test pyproject doesn't install pytest
-        dep_file_path = result.project.join(_DEPENDENCY_FILE)
-        lines = dep_file_path.readlines()
-        assert 'pytest = "*"\n' not in lines
-        # Test contents of test file
-        test_file_path = result.project.join("tests/test_app.py")
-        lines = test_file_path.readlines()
-        assert "import unittest" in "".join(lines)
-        assert "import pytest" not in "".join(lines)
-
-
-def test_docstrings_style(cookies):
-    with bake_in_temp_dir(
-        cookies, extra_context={"docstrings_style": "google"}
-    ) as result:
-        assert result.project.isdir()
         # Test lint rule contains google style
         flake8_conf_file_apth = result.project.join(".flake8")
         lines = flake8_conf_file_apth.readlines()
         assert "docstring-convention = google" in "".join(lines)
 
-
-# def test_project_with_hyphen_in_module_name(cookies):
-#     result = cookies.bake(
-#         extra_context={'project_name': 'something-with-a-dash'}
-#     )
-#     assert result.project is not None
-#     project_path = str(result.project)
-#
-#     # when:
-#     travis_setup_cmd = ('python travis_pypi_setup.py'
-#                         ' --repo audreyr/cookiecutter-pypackage'
-#                         ' --password invalidpass')
-#     run_inside_dir(travis_setup_cmd, project_path)
-#
-#     # then:
-#     result_travis_config = yaml.load(
-#         open(os.path.join(project_path, ".travis.yml"))
-#     )
-#     assert "secure" in result_travis_config["deploy"]["password"],\
-#         "missing password config in .travis.yml"
+        assert result.project.isdir()
+        # Test pyproject does install pytest
+        dep_file_path = result.project.join(_DEPENDENCY_FILE)
+        lines = dep_file_path.readlines()
+        assert 'pytest = "*"\n' in lines
+        # Test contents of test file
+        test_file_path = result.project.join("tests/test_app.py")
+        lines = test_file_path.readlines()
+        assert "import unittest" not in "".join(lines)
+        assert "import pytest" in "".join(lines)
 
 
 @pytest.mark.parametrize(