+++ {{cookiecutter.project_slug}}/repo.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+# TODO: Rework file to use Cruft
 
 # !!!NOTICE!!
 # Personal token with full access rights is required to run this scripts
@@ -6,23 +7,16 @@
 
 # Create repo and push code to github
 gh repo create {{cookiecutter.project_slug}} --public
-git remote add origin git@github.com:{{cookiecutter.github_username}}/{{cookiecutter.project_slug}}.git
+git remote add origin git@github.com:Loveall-Labs/{{cookiecutter.project_slug}}.git
 git add .
 pre-commit run --all-files
 git add .
 git commit -m "Initial commit by ppw"
 git branch -M main
 
-# Config github secret used by github workflow. 
+# Config github secret used by github workflow.
 gh secret set PERSONAL_TOKEN --body $GH_TOKEN
 gh secret set PYPI_API_TOKEN --body $PYPI_API_TOKEN
 gh secret set TEST_PYPI_API_TOKEN --body $TEST_PYPI_API_TOKEN
-
-# uncomment the following if you need to setup email notification
-# gh secret set BUILD_NOTIFY_MAIL_SERVER --body $BUILD_NOTIFY_MAIL_SERVER
-# gh secret set BUILD_NOTIFY_MAIL_PORT --body $BUILD_NOTIFY_MAIL_PORT
-# gh secret set BUILD_NOTIFY_MAIL_FROM --body $BUILD_NOTIFY_MAIL_FROM
-# gh secret set BUILD_NOTIFY_MAIL_PASSWORD --body $BUILD_NOTIFY_MAIL_PASSWORD
-# gh secret set BUILD_NOTIFY_MAIL_RCPT --body $BUILD_NOTIFY_MAIL_RCPT
 
 git push -u origin main