+++ {{cookiecutter.project_slug}}/{{cookiecutter.python_package_name}}/__init__.py
@@ -1,5 +1,5 @@
 """Top-level package for {{ cookiecutter.project_name }}."""
 
-__author__ = """{{ cookiecutter.full_name }}"""
-__email__ = "{{ cookiecutter.email }}"
+__author__ = """Loveall Labs LLC"""
+__email__ = "engineering@lovealllabs.com"
 __version__ = "{{ cookiecutter.version }}"