+++ {{cookiecutter.project_slug}}/{{cookiecutter.python_package_name}}/cli.py
@@ -1,12 +1,12 @@
-"""Console script for {{cookiecutter.project_slug}}."""
+"""Console script for {{cookiecutter.python_package_name}}."""
 
 {% if cookiecutter.command_line_interface|lower == 'fire' -%}
 import fire
 
 
 def help() -> None:
-    print("{{ cookiecutter.project_slug }}")
-    print("=" * len("{{ cookiecutter.project_slug }}"))
+    print("{{ cookiecutter.python_package_name }}")
+    print("=" * len("{{ cookiecutter.python_package_name }}"))
     print("{{ cookiecutter.project_short_description }}")
 
 def main() -> None: