+++ {{cookiecutter.project_slug}}/HISTORY.md
@@ -2,4 +2,4 @@
 
 ## {{ cookiecutter.version }} ({% now 'local' %})
 
-* First release on PyPI.
+* Project created.